--- page.js
@@ -1,6 +1,0 @@
-
-import { redirect } from 'next/navigation';
-
-export default function Home() {
-  redirect('/login');
-}
