+++ src/app/Principal/layout.js
@@ -1,0 +1,27 @@
+import { Inter } from 'next/font/google';
+import '../globals.css';
+import Footer from '../components/footer/footer';
+import Header from '../components/header/header';
+
+
+const inter = Inter({
+  subsets: ['latin'],
+  variable: '--font-inter',
+});
+
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({ children }) {
+  return (
+    <html lang="es">
+      <body className={inter.variable}>
+        <Header />
+        {children}
+        <Footer />
+      </body>
+    </html>
+  );
+}
