+++ src/app/layout.js
@@ -1,7 +1,3 @@
-import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-import Footer from "./components/footer/footer";
-import Header from "./components/header/header";
 import { Inter } from 'next/font/google';
 
 const inter = Inter({
@@ -9,30 +5,11 @@ const inter = Inter({
   variable: '--font-inter',
 });
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({ children }) {
+export default function LoginLayout({ children }) {
   return (
-    <html lang="en">
-      <body
-        className={`${inter.variable} ${geistSans.variable} ${geistMono.variable} antialiased`} /*gracias poderoso gpt por escribir esto que no se que es*/
-      >
-        <Header />
+    <html lang="es">
+      <body className={`${inter.variable}`}>
         {children}
-        <Footer />
       </body>
     </html>
   );
