--- src/app/login/layout.js
@@ -1,16 +1,0 @@
-import { Inter } from 'next/font/google';
-
-const inter = Inter({
-  subsets: ['latin'],
-  variable: '--font-inter',
-});
-
-export default function LoginLayout({ children }) {
-  return (
-    <html lang="es">
-      <body className={inter.variable}>
-        {children}
-      </body>
-    </html>
-  );
-}
